+++ lib/src/services/ssd_cneos/sbdb_cad_api/data/sbdb_cad_body.dart
@@ -11,7 +11,8 @@ class SbdbCadBody with _$SbdbCadBody {
   const factory SbdbCadBody.two00({
     required Signature signature,
     required int count,
-    required List<SbdbCadData>? data,
+    List<SbdbCadData>? data,
+    int? total,
     required JsonMap? rawBody,
   }) = _SbdbCadBody;